# docs/swagger.yaml
openapi: 3.0.3
info:
  title: FanFare API
  description: |
    Pop Culture Prediction & Sentiment Platform Backend API.
    
    FanFare enables fans to predict outcomes, track community sentiment,
    contribute crowdsourced context cards, and build reputation through
    accurate insights — not follower counts.
    
    ## 4 Key Product Pillars
    1. **Predictions Drive Engagement** — Users predict box office, awards, ratings, story events
    2. **Sentiment Dashboards Show the Fandom Pulse** — Track hype, satisfaction, controversies over time
    3. **Context Cards Reduce Misinformation** — Structured crowd-sourced notes clarify truth vs rumor
    4. **Reputation from Insight, Not Followers** — Points for accurate predictions & helpful contributions
  version: 1.0.0
  contact:
    name: FanFare Team
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
  schemas:
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          example: "lore_master_jane"
        email:
          type: string
          format: email
          example: "jane@example.com"
        password:
          type: string
          minLength: 6
          example: "securepass123"
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "lore_master_jane"
        password:
          type: string
          example: "securepass123"
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Signed JWT token
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatar_url:
          type: string
        bio:
          type: string
        reputation_score:
          type: integer
        followed_universes:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserProfileStats:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        predictions_attempted:
          type: integer
        predictions_correct:
          type: integer
        prediction_accuracy:
          type: number
          format: float
        context_cards_added:
          type: integer
        context_net_helpful:
          type: integer
        polls_voted:
          type: integer
        badges:
          type: array
          items:
            type: string
          example: ["Analyst", "Lore Master"]
...
    FollowFandomRequest:
      type: object
      required: [universe_id]
      properties:
        universe_id:
          type: string
    LeaderboardEntry:
      type: object
      properties:
        username:
          type: string
        reputation_score:
          type: integer
        rank:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register a new user account
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/login:
    post:
      summary: Authenticate and obtain a JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/users/me:
    get:
      summary: Get the currently authenticated user's own profile
      description: Returns the full profile and stats of the token holder — no need to know your own username.
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Authenticated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileStats'
        '401':
          description: Unauthorized — valid JWT required
  /api/v1/users/me/reputation:
    get:
      summary: Get the authenticated user's reputation audit trail
      description: Returns a chronological list of every point-earning action (prediction wins, context contributions, poll votes).
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Reputation history list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    action_type:
                      type: string
                      example: "prediction_correct"
                    points:
                      type: integer
                    reference_id:
                      type: string
                    created_at:
                      type: string
                      format: date-time
        '401':
          description: Unauthorized — valid JWT required
  /api/v1/users/{username}:
    get:
      summary: Get user profile with stats and badges
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile with dynamic stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileStats'
        '401':
          description: Unauthorized — valid JWT required
        '404':
          description: User not found
  /api/v1/users/follow-fandom:
    post:
      summary: Follow a franchise/universe
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowFandomRequest'
      responses:
        '200':
          description: Successfully followed
        '401':
          description: Unauthorized
  /api/v1/scoreboard:
    get:
      summary: Get global or fandom leaderboard
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: universe_id
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
        '401':
          description: Unauthorized — valid JWT required
...